docs and fixing circle
