+++ src/Command.js
@@ -1,7 +1,16 @@
-const Arguments = require('./Arguments');
+const Arguments = require('./Argument');
 const Http = require('./Http');
 
+/**
+ * Describes a command.
+ */
 class Command {
+  /**
+   * Build a {@link Command}.
+   *
+   * @param {String} name The given name
+   * @param {Object} rawCommand The raw data
+   */
   constructor(name, rawCommand) {
     this._name = name;
     this._output = rawCommand.output;
@@ -28,24 +37,50 @@ class Command {
     }
   }
 
-  get name () {
+  /**
+   * Get's the name of this {@link Command}.
+   *
+   * @return {String} The name
+   */
+  get name() {
     return this._name;
   }
 
+  /**
+   * The output type of this {@link Command}.
+   *
+   * @return {String} The output type
+   */
   get output() {
     return this._output;
   }
 
+  /**
+   * Get this {@Command}'s help.
+   *
+   * @return {String|null} The help
+   */
   get help() {
     return this._help;
   }
 
+  /**
+   * Get the format of this {@link Command}.
+   *
+   * @return {String|null}
+   */
   get format() {
     return this._format;
   }
 
+  /**
+   * Checks if the required arguments are supplied.
+   *
+   * @param {Object} _arguments The given argument mapping
+   * @return {Boolean} True if all required arguments are supplied, otherwise false
+   */
   areRequiredArgumentsSupplied(_arguments) {
-    const requiredArguments = this.arguments.filter(argument => argument.isRequired()).map(argument => argument.name);
+    const requiredArguments = this.arguments.filter((argument) => argument.isRequired()).map((argument) => argument.name);
     for (let i = 0; i < requiredArguments.length; i += 1) {
       if (!Object.keys(_arguments).includes(requiredArguments[i])) {
         return false;
@@ -55,8 +90,9 @@ class Command {
   }
 
   /**
+   * Get the {@ink Arguments}s for this {@link Command}.
    *
-   * @return {Array<Arguments>}
+   * @return {Array<Arguments>} The {@link Arguments}s
    */
   get arguments() {
     if (this._argumentsMap === null) {
@@ -65,6 +101,13 @@ class Command {
     return Object.values(this._argumentsMap);
   }
 
+  /**
+   * Get an {@link Argument} based on given argument from this {@link Command}.
+   *
+   * @param {String} argument The given argument
+   * @throws {String} If the argument does not exists
+   * @return {Command} The command with given name
+   */
   getArgument(argument) {
     if ((this._argumentsMap === null) || (!this._argumentsMap[argument])) {
       throw 'Argument does not exist';
@@ -73,14 +116,16 @@ class Command {
   }
 
   /**
+   * The this {@link Command}s {@link Http} service.
    *
-   * @returns {Http}
+   * @return {Http} The {@link Http} service
    */
   get http() {
     return this._http;
   }
 
   /**
+   * The the run object for this {@link Command}
    *
    * @return {null|{command: String, args: String, port: Number}}
    */
